1차 분리
