+++ server/ssts/src/main/java/com/ssts/ssts/domain/vote/repository/VoteRepository.java
@@ -1,0 +1,7 @@
+package com.ssts.ssts.domain.vote.repository;
+
+import com.ssts.ssts.domain.vote.entity.Vote;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface VoteRepository extends JpaRepository<Vote, Long> {
+}
